bots/requirements.txt
matrix-nio[e2e]>=0.24.0
aiohttp>=3.9.0
pyyaml>=6.0
python-olm>=3.2.0
aiofiles>=23.0
